% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAS.R
\name{cal_all_tools}
\alias{cal_all_tools}
\title{calculation seven tools}
\usage{
cal_all_tools(
  seurat_object,
  gSets_path,
  tools = c("AUCell", "Vision", "GSVA", "ssGSEA", "plage", "zscore"),
  normalize = c("log", "CLR", "RC", "scran", "none"),
  mat_type = "counts",
  n_cores = 3,
  rand_seed = 123
)
}
\arguments{
\item{gSets_path}{patwhays/gene sets in clasical GMT format}

\item{normalize}{normalization method}

\item{counts}{gene expresion matrix, rows are genes and cols are cells}

\item{tool}{select PAS tools}

\item{filter}{whether filtering for genes expressed in less than 5% cells}
}
\description{
parameters `counts`,  `gSets_path` and `gSets` are consistent across all tools
}
\details{
other parameters see details of tools
}
