% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cal_fscLVM}
\alias{cal_fscLVM}
\title{calculate PAS by nine tools}
\usage{
cal_fscLVM(
  counts,
  gSets_path,
  type = "counts",
  n_hidden = 1,
  ifLog = T,
  rand_seed = 123
)
}
\description{
parameters `counts`,  `gSets_path` and `gSets` are consistent across all tools
}
\details{
other parameters see details of tools
}
