% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAS.R
\name{cal_PAS}
\alias{cal_PAS}
\title{calculation Pathway Activity Score}
\usage{
cal_PAS(
  seurat_object,
  tool = c("AUCell", "Vision", "GSVA", "ssGSEA", "plage", "zscore"),
  species = "none",
  pathway = "none",
  gmt_file = "none",
  normalize = "sctransform",
  n_cores = 3,
  rand_seed = 123
)
}
\arguments{
\item{tool}{tool name}

\item{species}{species; human or mouse}

\item{pathway}{abbreviation for pathway database}

\item{gmt_file}{pathways in GMT format}

\item{normalize}{normalization method}

\item{counts}{gene expresion matrix, rows are genes and cols are cells}

\item{filter}{whether filtering for genes expressed in less than 5 percent cells}
}
\description{
parameters `counts`,  `gSets_path` and `gSets` are consistent across all tools
}
\details{
other parameters see details of tools
}
