% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucells.R
\name{calAUC}
\alias{calAUC}
\title{AUCell functions}
\usage{
calAUC(data, gSets)
}
\arguments{
\item{data}{expression matrix or data.frame with row in genes and column in cells}

\item{gSets}{gene sets in list format}
}
\value{
AUC score matrix
}
\description{
calculate PAS by AUCell function
}
