% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.R
\name{calGsva}
\alias{calGsva}
\title{GSVA functions}
\usage{
calGsva(data, gSets, alpha = 0.25, rnaseq = F)
}
\arguments{
\item{data}{expression matrix or data.frame with row in genes and column in cells}

\item{gSets}{gene sets in list format}
}
\value{
gsva score matrix
}
\description{
GSVA function
}
