% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plage.R
\name{calPlage}
\alias{calPlage}
\title{plage functions}
\usage{
calPlage(data, gSets)
}
\arguments{
\item{data}{expression matrix or data.frame with row in genes and column in cells}

\item{gSets}{gene sets in list format}
}
\value{
plage score matrix
}
\description{
plage function
}
