% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.R
\name{calSSgsea}
\alias{calSSgsea}
\title{ssGSEA functions}
\usage{
calSSgsea(data, gSets, alpha = 0.25, normalization = T)
}
\arguments{
\item{data}{expression matrix or data.frame with row in genes and column in cells}

\item{gSets}{gene sets in list format}
}
\value{
ssgsea score matrix
}
\description{
ssGSEA function
}
