% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision.R
\name{calVsion}
\alias{calVsion}
\title{VISION functions}
\usage{
calVsion(data, gSets)
}
\arguments{
\item{data}{expression matrix or data.frame with row in genes and column in cells}

\item{gSets}{gene sets in list format}
}
\value{
vision score matrix
}
\description{
calculate PAS by Vision function
}
