% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{calZscore}
\alias{calZscore}
\title{zscore functions}
\usage{
calZscore(data, gSets)
}
\arguments{
\item{data}{expression matrix or data.frame with row in genes and column in cells}

\item{gSets}{gene sets in list format}
}
\value{
Z score matrix
}
\description{
zscore function
}
