% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGMT}
\alias{getGMT}
\title{functions to convert data types}
\usage{
getGMT(gmt_file_path, n_gene_thre = 0)
}
\arguments{
\item{gmt_file_path}{.gmt file path}

\item{n_gene_thre}{minimum number of genes in gene sets}
}
\description{
Reading a .gmt file containing pathways or gene sets
}
\details{
Input a .gmt file path and return a list which including gene sets
}
