% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{getPathways}
\alias{getPathways}
\title{loading pathways or gene sets}
\usage{
getPathways(species, pathway)
}
\arguments{
\item{species}{species; human or mouse}

\item{pathway}{abbreviation for pathway database
detials see https://github.com/zgyaru/PASBench}
}
\value{
a list containing pathways or gene sets
}
\description{
loading pathways or gene sets
}
