% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigsToSparseMatrix}
\alias{sigsToSparseMatrix}
\title{Utility method to load signatures into a sparse matrix}
\usage{
sigsToSparseMatrix(gSets, expression)
}
\arguments{
\item{gSets}{List of gene sets}

\item{expression}{numeric Matrix Genes x Cells}
}
\value{
sparseMatrix containing signature matched values
}
\description{
Utility method to load signatures into a sparse matrix
}
